#185 Make the copy button available on book details pages
