+++ src/ts/content/extensions/copy/index.ts
@@ -1,2 +1,2 @@
-import "./onSearch";
+import "./onView";
 import "./onEdit";
